delete_post_controller tests created

+++ src/business/post/delete_post/delete_post_controller.ts
@@ -1,26 +1,23 @@
-import { Router, Request } from "express";
-import { PostRepository } from "@db/post/post_repository.js";
-import { makeDeletePostService } from "business/post/delete_post/delete_post_service.js"
-import { PostgressDBError, PostError, UserError } from "@src/errors.js";
+import { Router, Request } from "express"
+import { PostRepository } from "@db/post/post_repository.js"
+import { makeDeletePostService } from "@business/post/delete_post/delete_post_service.js"
+import { isExpressError, ExpressError } from "@src/errors.js"
+import { userIdExists } from "@business/post/delete_post/delete_post_controller_aux.js"
 
 export function makeDeletePostRouter(repo: PostRepository) {
   const deletePostService = makeDeletePostService(repo)
 
   return Router().delete("/:id", async (req: Request<{ id: number }, object, object>, res) => {
     const { id } = req.params
-    const userId = req.userId
     try {
-      if (!userId) {
-        throw new UserError(403, makeDeletePostRouter, "User id not found")
-      }
-      await deletePostService.deletePost({ id, userId })
+      await deletePostService.deletePost({ id, userId: userIdExists(req.userId) })
       res.json({ message: "Post deleted" })
     } catch (error) {
-      if (error instanceof PostgressDBError || error instanceof PostError) {
-        const { statusCode, name, func, message } = error
-        res.status(statusCode).json({ name, func, message })
+      if (isExpressError(error as Error)) {
+        res.status((error as ExpressError).statusCode).json({ message: (error as Error).message })
+      } else {
+        res.status(500).json({ error: (error as Error).message })
       }
-      res.json({ error })
     }
   })
 }
